refactor(tui): f / p を normal_keybindings.py に移す
ファイルの分割基準がモードなのに、normal モードのキーである f / p だけを
modal_keybindings.py に置いていた。基準を揃えて normal 側へ移す。

f が呼ぶ open 関数は register 内にネストしていたため、各 filter_modal.py に
open_filter_modal() として出す。project_modal.py が open_project_modal() を
持つのと同じ形になり、3 つの modal が揃う。
