+++ src/redi/tui/keybindings/modal_keybindings.py
@@ -1,22 +1,13 @@
-"""modal 表示中のキーバインド。
-
-modal を開く `f` / `p` も、開いたあとの操作と近い場所に置きたいのでここで登録する。
-"""
+"""modal 表示中のキーバインド。"""
 
 from prompt_toolkit.key_binding import KeyBindings
 
-from redi.tui.choices import (
-    build_assignee_choices,
-    build_status_choices,
-    build_user_choices,
-)
 from redi.tui.issue.issue_tab import reload_with_filter
 from redi.tui.keybindings.keybinding_conditions import (
     Conditions,
-    clear_temporary_state,
     reset_preview_scroll,
 )
-from redi.tui.project_modal import apply_project_switch, open_project_modal
+from redi.tui.project_modal import apply_project_switch
 from redi.tui.state import IssueFilter, TimeEntryFilter, TuiState
 from redi.tui.time_entry.time_entry_tab import (
     reload_with_filter as time_entry_reload_with_filter,
@@ -24,7 +15,6 @@ from redi.tui.time_entry.time_entry_tab import (
 
 
 def register(kb: KeyBindings, state: TuiState, conditions: Conditions) -> None:
-    normal_mode = conditions.normal
     show_help_modal = conditions.help_modal
     show_filter_modal = conditions.issue_filter_modal
     show_time_entry_filter_modal = conditions.time_entry_filter_modal
@@ -38,50 +28,6 @@ def register(kb: KeyBindings, state: TuiState, conditions: Conditions) -> None:
     @kb.add("q", filter=show_error_modal)
     def _(event):
         state.error_modal = None
-
-    def _open_filter_modal() -> None:
-        modal = state.issue_tab.filter_modal
-        modal.status_choices = build_status_choices()
-        modal.assignee_choices = build_assignee_choices(
-            state.effective_project_id(), state.me_id
-        )
-        modal.status_cursor = 0
-        for idx, (api_val, _label) in enumerate(modal.status_choices):
-            if api_val == state.issue_tab.filter.status_id:
-                modal.status_cursor = idx
-                break
-        modal.assignee_cursor = 0
-        for idx, (api_val, _label) in enumerate(modal.assignee_choices):
-            if api_val == state.issue_tab.filter.assigned_to_id:
-                modal.assignee_cursor = idx
-                break
-        modal.focus = "status"
-        modal.show = True
-
-    def _open_time_entry_filter_modal() -> None:
-        modal = state.time_entry_tab.filter_modal
-        modal.user_choices = build_user_choices(
-            state.effective_project_id(), state.me_id
-        )
-        modal.user_cursor = 0
-        for idx, (api_val, _label) in enumerate(modal.user_choices):
-            if api_val == state.time_entry_tab.filter.user_id:
-                modal.user_cursor = idx
-                break
-        modal.show = True
-
-    @kb.add("f", filter=normal_mode)
-    def _(event):
-        clear_temporary_state(state)
-        if state.tab == "issues":
-            _open_filter_modal()
-        elif state.tab == "time_entries":
-            _open_time_entry_filter_modal()
-
-    @kb.add("p", filter=normal_mode)
-    def _(event):
-        clear_temporary_state(state)
-        open_project_modal(state)
 
     @kb.add("j", filter=show_project_modal)
     @kb.add("down", filter=show_project_modal)
